refactor: reorganize project structure and update imports for project tree components
