+++ apps/web/src/common/modules/page-layouts/type.ts
@@ -1,7 +1,7 @@
 import type { TranslateResult } from 'vue-i18n';
 import type { Location } from 'vue-router';
 
-import type { ProjectGroupTreeNodeData } from '@/services/project/v-shared/types/project-tree-type';
+import type { ProjectGroupTreeNodeData } from '@/common/modules/project/project-tree-type';
 
 export interface Breadcrumb {
     name: TranslateResult|string;
