+++ apps/web/src/services/project/v1/composables/use-project-tree-data.ts
@@ -7,13 +7,14 @@ import type { Location } from 'vue-router';
 
 import { get } from 'lodash';
 
+import type { TreeDisplayMap, TreeNode } from '@cloudforet/mirinae/types/data-display/tree/tree-view/type';
+
 import { useAllReferenceStore } from '@/store/reference/all-reference-store';
 import type { ProjectGroupReferenceMap } from '@/store/reference/project-group-reference-store';
 import type { ProjectReferenceMap } from '@/store/reference/project-reference-store';
 
-import { useProjectTreeStore } from '@/services/project/v-shared/stores/project-tree-store';
-import type { TreeNode, TreeDisplayMap } from '@/services/project/v-shared/tree/type';
 import { PROJECT_ROUTE_V1 } from '@/services/project/v1/routes/route-constant';
+import { useProjectTreeStore } from '@/services/project/v1/stores/project-tree-store';
 
 interface ParentGroupItem {
     name: string;
